@@ -1,4 +1,3 @@
-require "spec_helper"
 require_relative "../../lib/octopolo/convenience_wrappers"
 require_relative "../../lib/octopolo/git"
 
@@ -12,68 +11,68 @@ module Octopolo
 
   describe CLIWrapper do
     let(:foo) { Foo.new }
-    let(:cli) { stub(:CLI) }
+    let(:cli) { stub_const('CLI', class_double(CLI)) }
 
     context "#cli" do
       it "wraps around the CLI class" do
-        foo.cli.should == CLI
+        foo.cli.should eq(CLI)
       end
 
       it "uses the given CLI class" do
         foo.cli = cli
-        foo.cli.should == cli
+        foo.cli.should eq(cli)
       end
     end
   end
 
   describe ConfigWrapper do
     let(:foo) { Foo.new }
-    let(:config) { stub(:config) }
+    let(:config) { double(:config) }
 
     context "#config" do
       it "parses the current config" do
         Octopolo.should_receive(:config) { config }
-        foo.config.should == config
+        foo.config.should eq(config)
       end
 
       it "uses the given parsed config" do
         Config.should_not_receive(:parse)
         foo.config = config
-        foo.config.should == config
+        foo.config.should eq(config)
       end
     end
   end
 
   describe UserConfigWrapper do
     let(:foo) { Foo.new }
-    let(:user_config) { stub(:user_config) }
+    let(:user_config) { double(:user_config) }
 
     context "#user_config" do
       it "parses the current user config" do
         UserConfig.should_receive(:parse) { user_config }
-        foo.user_config.should == user_config
+        foo.user_config.should eq(user_config)
       end
 
       it "uses the given parsed config" do
         foo.user_config = user_config
         UserConfig.should_not_receive(:parse)
-        foo.user_config.should == user_config
+        foo.user_config.should eq(user_config)
       end
     end
   end
 
   describe GitWrapper do
     subject { Foo.new }
-    let(:git) { stub(:Git) }
+    let(:git) { stub_const('Git', class_double(Git)) }
 
     context "#git" do
       it "wraps around the Git class" do
-        subject.git.should == Git
+        subject.git.should eq(Git)
       end
 
       it "uses the given Git" do
         subject.git = git
-        subject.git.should == git
+        subject.git.should eq(git)
       end
     end
   end
