@@ -1,9 +1,8 @@
-require "spec_helper"
 require_relative "../../lib/octopolo/git"
 
 module Octopolo
   describe Git do
-    let(:cli) { stub(:CLI) }
+    let(:cli) { double(:CLI) }
 
     context ".perform(subcommand)" do
       let(:command) { "status" }
@@ -35,7 +34,7 @@ module Octopolo
 
       it "performs a command to filter current branch from list of branches" do
         cli.should_receive(:perform_quietly).with("git branch | grep '^* ' | cut -c 3-") { output }
-        Git.current_branch.should == name
+        Git.current_branch.should eq(name)
       end
 
       it "raises NotOnBranch if not on a branch" do
@@ -45,24 +44,24 @@ module Octopolo
 
       it "staging and deploy should be reserved branches" do
         Git.stub(:current_branch).and_return "staging.05.12"
-        Git.reserved_branch?.should be_true
+        Git.reserved_branch?.should be true
 
         Git.stub(:current_branch).and_return "deployable.05.12"
-        Git.reserved_branch?.should be_true
+        Git.reserved_branch?.should be true
 
         Git.stub(:current_branch).and_return "qaready.05.12"
-        Git.reserved_branch?.should be_true
+        Git.reserved_branch?.should be true
       end
 
       it "other branches should not be reserved branches" do
         Git.stub(:current_branch).and_return "not_staging.05.12"
-        Git.reserved_branch?.should_not be_true
+        Git.reserved_branch?.should_not be true
 
         Git.stub(:current_branch).and_return "not_deployable.05.12"
-        Git.reserved_branch?.should_not be_true
+        Git.reserved_branch?.should_not be true
 
         Git.stub(:current_branch).and_return "not_qaready.05.12"
-        Git.reserved_branch?.should_not be_true
+        Git.reserved_branch?.should_not be true
       end
     end
 
@@ -185,7 +184,7 @@ module Octopolo
         cli.should_receive(:say).with(" ")
         Git.should_receive(:perform).with("status")
 
-        expect{Git.alert_dirty_index message}.to raise_error
+        expect{Git.alert_dirty_index message}.to raise_error(Octopolo::Git::DirtyIndex)
       end
     end
 
@@ -249,7 +248,7 @@ module Octopolo
       it "prunes the remote branch list and grabs all the branch names" do
         Git.should_receive(:fetch)
         Git.should_receive(:perform_quietly).with("branch --remote") { raw_output }
-        Git.remote_branches.should == cleaned_names.sort
+        Git.remote_branches.should eq(cleaned_names.sort)
       end
     end
 
@@ -268,14 +267,14 @@ module Octopolo
         deployables = Git.branches_for(Git::DEPLOYABLE_PREFIX)
         deployables.should include depl1
         deployables.should include depl2
-        deployables.should == [depl1, depl2].sort
-        deployables.count.should == 2
+        deployables.should eq([depl1, depl2].sort)
+        deployables.count.should eq(2)
       end
 
       it "can find staging branches" do
         stagings = Git.branches_for(Git::STAGING_PREFIX)
         stagings.should include stage1
-        stagings.count.should == 1
+        stagings.count.should eq(1)
       end
     end
 
@@ -285,7 +284,7 @@ module Octopolo
 
       it "returns the last deployable branch" do
         Git.should_receive(:branches_for).with(Git::DEPLOYABLE_PREFIX) { [depl1, depl2] }
-        Git.deployable_branch.should == depl2
+        Git.deployable_branch.should eq(depl2)
       end
 
       it "raises an exception if none exist" do
@@ -300,7 +299,7 @@ module Octopolo
 
       it "returns the last staging branch" do
         Git.should_receive(:branches_for).with(Git::STAGING_PREFIX) { [stage1, stage2] }
-        Git.staging_branch.should == stage2
+        Git.staging_branch.should eq(stage2)
       end
 
       it "raises an exception if none exist" do
@@ -315,7 +314,7 @@ module Octopolo
 
       it "returns the last qaready branch" do
         Git.should_receive(:branches_for).with(Git::QAREADY_PREFIX) { [qaready1, qaready2] }
-        Git.qaready_branch.should == qaready2
+        Git.qaready_branch.should eq(qaready2)
       end
 
       it "raises an exception if none exist" do
@@ -345,8 +344,8 @@ module Octopolo
       it "returns the last #{Git::RECENT_TAG_LIMIT} tags" do
         Git.should_receive(:release_tags) { long_list }
         tags = Git.recent_release_tags
-        tags.count.should == Git::RECENT_TAG_LIMIT
-        tags.should == long_list.last(Git::RECENT_TAG_LIMIT)
+        tags.count.should eq(Git::RECENT_TAG_LIMIT)
+        tags.should eq(long_list.last(Git::RECENT_TAG_LIMIT))
       end
     end
 
@@ -393,7 +392,7 @@ module Octopolo
 
     context ".stale_branches(destination_branch, branches_to_ignore)" do
       let(:ignored) { %w(foo bar) }
-      let(:branch_name) { "master" }
+      let(:branch_name) { "main" }
       let(:sha) { "asdf123" }
       let(:raw_result) do
         %Q(
@@ -411,10 +410,10 @@ module Octopolo
         expect(Git.stale_branches(branch_name, ignored)).to eq(%w(bing bang))
       end
 
-      it "defaults to master branch and no extra branches to ignore" do
+      it "defaults to main branch and no extra branches to ignore" do
         Git.should_receive(:fetch)
         Git.should_receive(:stale_branches_to_ignore).with([]) { ignored }
-        Git.should_receive(:recent_sha).with("master") { sha }
+        Git.should_receive(:recent_sha).with("main") { sha }
         Git.should_receive(:perform_quietly).with("branch --remote --merged #{sha} | grep -E -v '(foo|bar)'") { raw_result }
 
         Git.stale_branches
@@ -424,14 +423,14 @@ module Octopolo
     context "#branches_to_ignore(custom_branch_list)" do
       it "ignores some branches by default" do
         expect(Git.send(:stale_branches_to_ignore)).to include "HEAD"
-        expect(Git.send(:stale_branches_to_ignore)).to include "master"
+        expect(Git.send(:stale_branches_to_ignore)).to include "main"
         expect(Git.send(:stale_branches_to_ignore)).to include "staging"
         expect(Git.send(:stale_branches_to_ignore)).to include "deployable"
       end
 
       it "accepts an optional list of additional branches to ignore" do
         expect(Git.send(:stale_branches_to_ignore, ["foo"])).to include "HEAD"
-        expect(Git.send(:stale_branches_to_ignore, ["foo"])).to include "master"
+        expect(Git.send(:stale_branches_to_ignore, ["foo"])).to include "main"
         expect(Git.send(:stale_branches_to_ignore, ["foo"])).to include "staging"
         expect(Git.send(:stale_branches_to_ignore, ["foo"])).to include "deployable"
         expect(Git.send(:stale_branches_to_ignore, ["foo"])).to include "foo"
