@@ -1,4 +1,3 @@
-require "spec_helper"
 require_relative "../../../lib/octopolo/github/issue_creator"
 
 module Octopolo
@@ -9,30 +8,29 @@ module Octopolo
       let(:options) { {} }
       let(:title) { "title" }
       let(:body) { "body" }
-      let(:pivotal_ids) { %w(123 456) }
       let(:jira_ids) { %w(123 456) }
       let(:jira_url) { "https://example-jira.com" }
 
       context ".perform repo_name, options" do
-        let(:creator) { stub }
+        let(:creator) { double }
 
         it "instantiates a creator and perfoms it" do
           IssueCreator.should_receive(:new).with(repo_name, options) { creator }
           creator.should_receive(:perform)
-          IssueCreator.perform(repo_name, options).should == creator
+          IssueCreator.perform(repo_name, options).should eq(creator)
         end
       end
 
       context ".new repo_name, options" do
         it "remembers the repo name and options" do
           creator = IssueCreator.new repo_name, options
-          creator.repo_name.should == repo_name
-          creator.options.should == options
+          creator.repo_name.should eq(repo_name)
+          creator.options.should eq(options)
         end
       end
 
       context "#perform" do
-        let(:data) { stub(:mash, number: 123) }
+        let(:data) { double(:mash, number: 123) }
 
         before do
           creator.stub({
@@ -43,9 +41,9 @@ module Octopolo
 
         it "generates the issue with the given details and retains the information" do
           GitHub.should_receive(:create_issue).with(repo_name, title, body, labels: []) { data }
-          creator.perform.should == data
-          creator.number.should == data.number
-          creator.data.should == data
+          creator.perform.should eq(data)
+          creator.number.should eq(data.number)
+          creator.data.should eq(data)
         end
 
         it "raises CannotCreate if any exception occurs" do
@@ -59,7 +57,7 @@ module Octopolo
 
         it "returns the stored issue number" do
           creator.number = number
-          creator.number.should == number
+          creator.number.should eq(number)
         end
 
         it "raises an exception if no issue has been created yet" do
@@ -69,11 +67,11 @@ module Octopolo
       end
 
       context "#data" do
-        let(:details) { stub(:data) }
+        let(:details) { double(:data) }
 
         it "returns the stored issue details" do
           creator.data = details
-          creator.data.should == details
+          creator.data.should eq(details)
         end
 
         it "raises an exception if no information has been captured yet" do
@@ -87,7 +85,7 @@ module Octopolo
           before { creator.options[:title] = title }
 
           it "fetches from the options" do
-            creator.title.should == title
+            creator.title.should eq(title)
           end
         end
 
@@ -97,40 +95,26 @@ module Octopolo
         end
       end
 
-      context "#pivotal_ids" do
-        it "fetches from the options" do
-          creator.options[:pivotal_ids] = pivotal_ids
-          creator.pivotal_ids.should == pivotal_ids
-        end
-
-        it "defaults to an empty array if it's missing" do
-          creator.options[:pivotal_ids] = nil
-          creator.pivotal_ids.should == []
-        end
-      end
-
       context "#body_locals" do
         let(:urls) { %w(link1 link2) }
 
         before do
           creator.stub({
-            pivotal_ids: pivotal_ids,
             jira_ids: jira_ids,
             jira_url: jira_url,
           })
         end
         it "includes the necessary keys to render the template" do
-          creator.body_locals[:pivotal_ids].should == creator.pivotal_ids
-          creator.body_locals[:jira_ids].should == creator.jira_ids
-          creator.body_locals[:jira_url].should == creator.jira_url
+          creator.body_locals[:jira_ids].should eq(creator.jira_ids)
+          creator.body_locals[:jira_url].should eq(creator.jira_url)
         end
       end
 
       context "#edit_body" do
-        let(:path) { stub(:path) }
-        let(:body) { stub(:string) }
-        let(:tempfile) { stub(:tempfile) }
-        let(:edited_body) { stub(:edited_body) }
+        let(:path) { double(:path) }
+        let(:body) { double(:string) }
+        let(:tempfile) { double(:tempfile) }
+        let(:edited_body) { double(:edited_body) }
 
         before do
           Tempfile.stub(:new) { tempfile }
@@ -144,7 +128,7 @@ module Octopolo
           end
 
           it "returns the un-edited output" do
-            creator.edit_body(body).should == body
+            creator.edit_body(body).should eq(body)
           end
         end
 
@@ -175,14 +159,14 @@ module Octopolo
           end
 
           it "returns the user edited output" do
-            creator.edit_body(body).should == edited_body
+            creator.edit_body(body).should eq(edited_body)
           end
         end
       end
 
       context "#body" do
-        let(:locals) { stub(:hash) }
-        let(:output) { stub(:string) }
+        let(:locals) { double(:hash) }
+        let(:output) { double(:string) }
 
         before do
           creator.stub({
@@ -192,11 +176,11 @@ module Octopolo
 
         it "renders the body template with the body locals" do
           Renderer.should_receive(:render).with(Renderer::ISSUE_BODY, locals) { output }
-          creator.body.should == output
+          creator.body.should eq(output)
         end
 
         context "when the editor option is set" do
-          let(:edited_output) { stub(:output) }
+          let(:edited_output) { double(:output) }
 
           before do
             creator.stub({
@@ -208,7 +192,7 @@ module Octopolo
           it "calls the edit_body method" do
             Renderer.should_receive(:render).with(Renderer::ISSUE_BODY, locals) { output }
             creator.should_receive(:edit_body).with(output) { edited_output }
-            creator.body.should == edited_output
+            creator.body.should eq(edited_output)
           end
         end
       end
