@@ -1,4 +1,3 @@
-require "spec_helper"
 require_relative "../../../lib/octopolo/github"
 
 module Octopolo
@@ -10,7 +9,7 @@ module Octopolo
       let(:labels_hash) { [label_hash_1,label_hash_2] }
       let(:label1) { Label.new(name: "low-risk", color: "343434") }
       let(:label2) { Label.new(name: "high-risk", color: '565656') }
-      let(:config) { stub(:config, github_repo: "foo") }
+      let(:config) { double(:config, github_repo: "foo") }
 
       subject { Label }
 
