@@ -1,4 +1,3 @@
-require "spec_helper"
 require_relative "../../../lib/octopolo/github"
 
 module Octopolo
@@ -6,21 +5,21 @@ module Octopolo
     describe PullRequest do
       let(:repo_name) { "account/repo" }
       let(:pr_number) { 7 }
-      let(:prhash) { stub }
-      let(:commits) { stub }
-      let(:comments) { stub }
-      let(:octo) { stub }
+      let(:prhash) { double }
+      let(:commits) { double }
+      let(:comments) { double }
+      let(:octo) { double }
 
       context ".new" do
         it "remembers the pull request identifiers" do
           pr = PullRequest.new repo_name, pr_number
-          pr.repo_name.should == repo_name
-          pr.number.should == pr_number
+          pr.repo_name.should eq(repo_name)
+          pr.number.should eq(pr_number)
         end
 
         it "optionally accepts the github data" do
           pr = PullRequest.new repo_name, pr_number, octo
-          pr.data.should == octo
+          pr.data.should eq(octo)
         end
 
         it "fails if not given a repo name" do
@@ -37,7 +36,7 @@ module Octopolo
 
         it "fetches the details from GitHub" do
           GitHub.should_receive(:pull_request).with(pull.repo_name, pull.number) { octo }
-          pull.data.should == octo
+          pull.data.should eq(octo)
         end
 
         it "catches the information" do
@@ -60,25 +59,25 @@ module Octopolo
         end
 
         context "#title" do
-          let(:octo) { stub(title: "the title") }
+          let(:octo) { double(title: "the title") }
 
           it "retrieves from the github data" do
-            pull.title.should == octo.title
+            pull.title.should eq(octo.title)
           end
         end
 
         context "#branch" do
-          let(:octo) { stub(head: stub(ref: "asdf")) }
+          let(:octo) { double(head: double(ref: "asdf")) }
 
           it "retrieves from the github data" do
-            pull.branch.should == octo.head.ref
+            pull.branch.should eq(octo.head.ref)
           end
         end
 
         context "#commits" do
           it "fetches through octokit" do
             Commit.should_receive(:for_pull_request).with(pull) { commits }
-            pull.commits.should == commits
+            pull.commits.should eq(commits)
           end
 
           it "caches the result" do
@@ -91,7 +90,7 @@ module Octopolo
         context "#comments" do
           it "fetches through octokit" do
             GitHub.should_receive(:issue_comments).with(pull.repo_name, pull.number) { comments }
-            pull.comments.should == comments
+            pull.comments.should eq(comments)
           end
 
           it "caches the result" do
@@ -102,8 +101,8 @@ module Octopolo
         end
 
         context "#author_names" do
-          let(:commit1) { stub(author_name: "foo") }
-          let(:commit2) { stub(author_name: "bar") }
+          let(:commit1) { double(author_name: "foo") }
+          let(:commit2) { double(author_name: "bar") }
 
           before do
             pull.stub(commits: [commit1, commit2])
@@ -112,46 +111,46 @@ module Octopolo
           it "returns the list of authors" do
             names = pull.author_names
             names.should_not be_empty
-            names.count.should == 2
-            names.first.should == "foo"
+            names.count.should eq(2)
+            names.first.should eq("foo")
           end
 
           it "returns only unique values" do
             # make it same commenter
             commit2.stub(author_name: commit1.author_name)
             names = pull.author_names
-            names.size.should == 1
+            names.size.should eq(1)
           end
         end
 
         context "#commenter_names" do
-          let(:comment1) { stub(user: stub(login: "pbyrne")) }
-          let(:comment2) { stub(user: stub(login: "anfleene")) }
+          let(:comment1) { double(user: double(login: "pbyrne")) }
+          let(:comment2) { double(user: double(login: "anfleene")) }
 
           before do
             pull.stub(comments: [comment1, comment2], author_names: [])
-            GitHub::User.stub(:new).with("pbyrne").and_return(stub(:author_name => "pbyrne"))
-            GitHub::User.stub(:new).with("anfleene").and_return(stub(:author_name => "anfleene"))
+            GitHub::User.stub(:new).with("pbyrne").and_return(double(:author_name => "pbyrne"))
+            GitHub::User.stub(:new).with("anfleene").and_return(double(:author_name => "anfleene"))
           end
 
           it "returns the names of the commit authors" do
             names = pull.commenter_names
             names.should_not be_empty
-            names.size.should == 2
-            names.first.should == "pbyrne"
+            names.size.should eq(2)
+            names.first.should eq("pbyrne")
           end
 
           it "returns only unique values" do
             # make it same commenter
             comment2.user.stub(login: comment1.user.login)
             names = pull.commenter_names
-            names.size.should == 1
+            names.size.should eq(1)
           end
 
           it "does not include authors in this list" do
             pull.stub(author_names: [comment1.user.login])
             names = pull.commenter_names
-            names.size.should == 1
+            names.size.should eq(1)
             names.should_not include comment1.user.login
           end
         end
@@ -166,10 +165,10 @@ module Octopolo
         end
 
         context "#url" do
-          let(:octo) { stub(html_url: "http://example.com") }
+          let(:octo) { double(html_url: "http://example.com") }
 
           it "retrieves from the github data" do
-            pull.url.should == octo.html_url
+            pull.url.should eq(octo.html_url)
           end
         end
 
@@ -193,7 +192,7 @@ module Octopolo
 
           it "parses from the body" do
             urls = pull.external_urls
-            urls.size.should == 3
+            urls.size.should eq(3)
             urls.should include "http://thedesk.tstmedia.com/admin.php?pg=request&reqid=44690"
             urls.should include "http://thedesk.tstmedia.com/admin.php?pg=request&reqid=44693"
             urls.should include "http://www.ngin.com.stage.ngin-staging.com/api/volleyball/stats/summaries?id=68382&gender=girls&tst_test=1&date=8/24/2012"
@@ -201,15 +200,15 @@ module Octopolo
         end
 
         context "#body" do
-          let(:octo) { stub(body: "asdf") }
+          let(:octo) { double(body: "asdf") }
 
           it "retrieves from the github data" do
-            pull.body.should == octo.body
+            pull.body.should eq(octo.body)
           end
 
           it "returns an empty string if the GitHub data has no body" do
             octo.stub(body: nil)
-            pull.body.should == ""
+            pull.body.should eq("")
           end
         end
 
@@ -225,7 +224,7 @@ module Octopolo
         context "#week" do
           it "retrieves from the github data" do
             octo.stub(closed_at: "2012-09-18T14:00:01Z")
-            pull.week.should == Week.parse(octo.closed_at)
+            pull.week.should eq(Week.parse(octo.closed_at))
           end
         end
       end
@@ -236,22 +235,22 @@ module Octopolo
 
         it "infers from the repo_name" do
           pull.repo_name = "account/foo"
-          pull.human_app_name.should == "Foo"
+          pull.human_app_name.should eq("Foo")
           pull.repo_name = "account/foo_bar"
-          pull.human_app_name.should == "Foo Bar"
+          pull.human_app_name.should eq("Foo Bar")
         end
       end
 
       context ".closed repo_name" do
-        let(:raw_pr) { stub(:octo_data, number: 123) }
-        let(:pr_wrapper) { stub(:pull_request) }
+        let(:raw_pr) { double(:octo_data, number: 123) }
+        let(:pr_wrapper) { double(:pull_request) }
 
         it "crawls the repo for pull requests and wraps them in PullRequests" do
           GitHub.should_receive(:pull_requests).with(repo_name, "closed") { [raw_pr] }
           PullRequest.should_receive(:new).with(repo_name, raw_pr.number, raw_pr) { pr_wrapper }
 
           result = PullRequest.closed(repo_name)
-          result.should == [pr_wrapper]
+          result.should eq([pr_wrapper])
         end
       end
 
@@ -274,16 +273,16 @@ module Octopolo
       end
 
       context ".create repo_name, options" do
-        let(:options) { stub(:hash) }
-        let(:number) { stub(:integer) }
-        let(:data) { stub(:data)}
-        let(:creator) { stub(:pull_request_creator, number: number, data: data)}
-        let(:pull_request) { stub(:pull_request) }
+        let(:options) { double(:hash) }
+        let(:number) { double(:integer) }
+        let(:data) { double(:data)}
+        let(:creator) { double(:pull_request_creator, number: number, data: data)}
+        let(:pull_request) { double(:pull_request) }
 
         it "passes on to PullRequestCreator and returns a new PullRequest" do
           PullRequestCreator.should_receive(:perform).with(repo_name, options) { creator }
           PullRequest.should_receive(:new).with(repo_name, number, data) { pull_request }
-          PullRequest.create(repo_name, options).should == pull_request
+          PullRequest.create(repo_name, options).should eq(pull_request)
         end
       end
 
@@ -300,34 +299,34 @@ module Octopolo
           Git.should_receive(:current_branch) { branch_name }
           CLI.should_receive(:say).with("Pull request for current branch is number #{pr_number}")
           GitHub.should_receive(:search_issues) { double(total_count: 1, items: [pull]) }
-          PullRequest.current.should == pull
+          PullRequest.current.should eq(pull)
         end
 
         it "returns nil when Git.current_branch fails" do
           Git.should_receive(:current_branch) { raise error_message }
           CLI.should_receive(:say).with("An error occurred while getting the current branch: #{error_message}")
-          PullRequest.current.should == nil
+          PullRequest.current.should eq(nil)
         end
 
         it "returns nil when GitHub.pull_requests fails" do
           Git.should_receive(:current_branch) { branch_name }
           GitHub.should_receive(:search_issues) { raise error_message }
           CLI.should_receive(:say).with("An error occurred while getting the current branch: #{error_message}")
-          PullRequest.current.should == nil
+          PullRequest.current.should eq(nil)
         end
 
         it "returns nil when more than one PR exists" do
           Git.should_receive(:current_branch) { branch_name }
           GitHub.should_receive(:search_issues) { double(total_count: 2, items: [pull, pull]) }
           CLI.should_receive(:say).with("Multiple pull requests found for branch #{branch_name}")
-          PullRequest.current.should == nil
+          PullRequest.current.should eq(nil)
         end
 
         it "returns nil when no PR exists" do
           Git.should_receive(:current_branch) { branch_name }
           GitHub.should_receive(:search_issues) { double(total_count: 0, items: []) }
           CLI.should_receive(:say).with("No pull request found for branch #{branch_name}")
-          PullRequest.current.should == nil
+          PullRequest.current.should eq(nil)
         end
 
       end
