@@ -1,4 +1,3 @@
-require "spec_helper"
 require_relative "../../../lib/octopolo/scripts/issue"
 require_relative "../../../lib/octopolo/github/issue"
 
@@ -6,17 +5,16 @@ module Octopolo
   module Scripts
     describe Issue do
       let(:config) do
-        stub(:config, {
+        double(:config, {
           deploy_branch: "production",
           github_repo: "tstmedia/foo",
-          use_pivotal_tracker: true,
           use_jira: true
         })
       end
-      let(:cli) { stub(:cli) }
-      let(:git) { stub(:Git) }
+      let(:cli) { double(:cli) }
+      let(:git) { double(:Git) }
       let(:issue_url) { "http://github.com/tstmedia/octopolo/issues/0" }
-      let(:issue) { stub(:issue) }
+      let(:issue) { double(:issue) }
 
       subject { Issue.new }
 
@@ -43,7 +41,6 @@ module Octopolo
           GitHub.should_receive(:connect).and_yield
           expect(subject).to receive(:ask_questionaire)
           expect(subject).to receive(:create_issue)
-          expect(subject).to receive(:update_pivotal)
           expect(subject).to receive(:update_jira)
           expect(subject).to receive(:update_labels)
           expect(subject).to receive(:open_in_browser)
@@ -61,7 +58,6 @@ module Octopolo
         it "asks appropriate questions to create a issue" do
           expect(subject).to receive(:announce)
           expect(subject).to receive(:ask_title)
-          expect(subject).to receive(:ask_pivotal_ids)
           expect(subject).to receive(:ask_jira_ids)
           expect(subject).to receive(:ask_labels)
 
@@ -105,31 +101,8 @@ module Octopolo
         end
       end
 
-      context "#ask_pivotal_ids" do
-        let(:ids_with_whitespace) { "123 456" }
-        let(:ids_with_commas) { "234, 567" }
-
-        it "asks for and captures IDs for related pivotal tasks" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { ids_with_whitespace }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq(%w(123 456))
-        end
-
-        it "asks for and captures IDs with commas" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { ids_with_commas }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq(%w(234 567))
-        end
-
-        it "sets to an empty array if not provided an ansswer" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { "" }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq([])
-        end
-      end
-
       context "#create_issue" do
-        let(:attributes) { stub(:hash) }
+        let(:attributes) { double(:hash) }
 
         before do
           subject.stub(:issue_attributes) { attributes }
@@ -145,16 +118,14 @@ module Octopolo
       context "#issue_attributes" do
         before do
           subject.title = "title"
-          subject.pivotal_ids = %w(123)
         end
 
         it "combines the anssers with a handful of deault values" do
-          subject.send(:issue_attributes).should == {
+          subject.send(:issue_attributes).should eq({
             title: subject.title,
-            pivotal_ids: subject.pivotal_ids,
             jira_ids: subject.jira_ids,
             editor: nil
-          }
+          })
         end
       end
 
@@ -169,39 +140,10 @@ module Octopolo
         end
       end
 
-      context "#update_pivotal" do
-        before do
-          subject.pivotal_ids = %w(123 234)
-          subject.issue = stub(url: "test")
-        end
-        let(:story_commenter) { stub(perform: true) }
-
-        it "creates a story commenter for each pivotal_id" do
-          Pivotal::StoryCommenter.should_receive(:new).with("123", "test") { story_commenter }
-          Pivotal::StoryCommenter.should_receive(:new).with("234", "test") { story_commenter }
-          subject.send(:update_pivotal)
-        end
-
-      end
-
-      context "#update_jira" do
-        before do
-          subject.jira_ids = %w(123 234)
-          subject.issue = stub(url: "test")
-        end
-        let(:story_commenter) { stub(perform: true) }
-
-        it "creates a story commenter for each pivotal_id" do
-          Jira::StoryCommenter.should_receive(:new).with("123", "test") { story_commenter }
-          Jira::StoryCommenter.should_receive(:new).with("234", "test") { story_commenter }
-          subject.send(:update_jira)
-        end
-      end
-
       context "#update_labels" do
         before do
           subject.labels = "high-risk"
-          subject.issue = stub()
+          subject.issue = double()
         end
         it "calls update_labels with proper arguments" do
           expect(subject.issue).to receive(:add_labels).with('high-risk')
