@@ -1,22 +1,20 @@
-require "spec_helper"
 require_relative "../../../lib/octopolo/scripts/pull_request"
 
 module Octopolo
   module Scripts
     describe PullRequest do
       let(:config) do
-        stub(:config, {
+        double(:config, {
           deploy_branch: "production",
           github_repo: "tstmedia/foo",
-          use_pivotal_tracker: true,
           use_jira: true
         })
       end
-      let(:cli) { stub(:cli) }
+      let(:cli) { double(:cli) }
       let(:current_branch) { "bug-123-something" }
-      let(:git) { stub(:Git, current_branch: current_branch, reserved_branch?: false) }
+      let(:git) { double(:Git, current_branch: current_branch, reserved_branch?: false) }
       let(:pull_request_url) { "http://github.com/tstmedia/octopolo/pull/0" }
-      let(:pull_request) { stub(:pull_request) }
+      let(:pull_request) { double(:pull_request) }
 
       subject { PullRequest.new }
 
@@ -47,7 +45,6 @@ module Octopolo
           GitHub.should_receive(:connect).and_yield
           expect(subject).to receive(:ask_questionnaire)
           expect(subject).to receive(:create_pull_request)
-          expect(subject).to receive(:update_pivotal)
           expect(subject).to receive(:update_jira)
           expect(subject).to receive(:update_labels)
           expect(subject).to receive(:open_in_browser)
@@ -65,7 +62,6 @@ module Octopolo
         it "asks appropriate questions to create a pull request" do
           expect(subject).to receive(:announce)
           expect(subject).to receive(:ask_title)
-          expect(subject).to receive(:ask_pivotal_ids)
           expect(subject).to receive(:ask_jira_ids)
           expect(subject).to receive(:ask_labels)
 
@@ -76,7 +72,6 @@ module Octopolo
           before do
             subject.stub(:announce)
             subject.stub(:ask_title)
-            subject.stub(:ask_pivotal_ids)
             subject.stub(:ask_jira_ids)
             subject.stub(:ask_labels)
           end
@@ -179,31 +174,8 @@ module Octopolo
         end
       end
 
-      context "#ask_pivotal_ids" do
-        let(:ids_with_whitespace) { "123 456" }
-        let(:ids_with_commas) { "234, 567" }
-
-        it "asks for and captures IDs for related pivotal tasks" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { ids_with_whitespace }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq(%w(123 456))
-        end
-
-        it "asks for and captures IDs with commas" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { ids_with_commas }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq(%w(234 567))
-        end
-
-        it "sets to an empty array if not provided an ansswer" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { "" }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq([])
-        end
-      end
-
       context "#create_pull_request" do
-        let(:attributes) { stub(:hash) }
+        let(:attributes) { double(:hash) }
 
         before do
           subject.stub(:pull_request_attributes) { attributes }
@@ -223,20 +195,18 @@ module Octopolo
       context "#pull_request_attributes" do
         before do
           subject.title = "title"
-          subject.destination_branch = "some-branch",
-          subject.pivotal_ids = %w(123)
+          subject.destination_branch = "some-branch"
         end
 
         it "combines the anssers with a handful of deault values" do
-          subject.send(:pull_request_attributes).should == {
+          subject.send(:pull_request_attributes).should eq({
             title: subject.title,
             destination_branch: subject.destination_branch,
             source_branch: git.current_branch,
-            pivotal_ids: subject.pivotal_ids,
             jira_ids: subject.jira_ids,
             editor: nil,
             skip_draft: nil
-          }
+          })
         end
       end
 
@@ -251,39 +221,10 @@ module Octopolo
         end
       end
 
-      context "#update_pivotal" do
-        before do
-          subject.pivotal_ids = %w(123 234)
-          subject.pull_request = stub(url: "test")
-        end
-        let(:story_commenter) { stub(perform: true) }
-
-        it "creates a story commenter for each pivotal_id" do
-          Pivotal::StoryCommenter.should_receive(:new).with("123", "test") { story_commenter }
-          Pivotal::StoryCommenter.should_receive(:new).with("234", "test") { story_commenter }
-          subject.send(:update_pivotal)
-        end
-
-      end
-
-      context "#update_jira" do
-        before do
-          subject.jira_ids = %w(123 234)
-          subject.pull_request = stub(url: "test")
-        end
-        let(:story_commenter) { stub(perform: true) }
-
-        it "creates a story commenter for each pivotal_id" do
-          Jira::StoryCommenter.should_receive(:new).with("123", "test") { story_commenter }
-          Jira::StoryCommenter.should_receive(:new).with("234", "test") { story_commenter }
-          subject.send(:update_jira)
-        end
-      end
-
       context "#update_labels" do
         before do
           subject.labels = ["high-risk"]
-          subject.pull_request = stub()
+          subject.pull_request = double()
         end
         it "calls update_labels with proper arguments" do
           expect(subject.pull_request).to receive(:add_labels).with(['high-risk'])
